apply with_thread_limits() to HPUModelRunner.load_model() only
